% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annovar.R
\name{annovar_annotate_region}
\alias{annovar_annotate_region}
\title{Annotate variant regions}
\usage{
annovar_annotate_region(avinput.file, output.prefix,
  db.path = "/imbs/external_data/annotation_and_references/annovar/humandb",
  build = "hg19", exec = "annotate_variation.pl")
}
\arguments{
\item{avinput.file}{[\code{string}]\cr
The file name of the ANNOVAR input file.}

\item{output.prefix}{[\code{string}]\cr
The basename of the output files.}

\item{db.path}{[\code{string}]\cr
The path to the database files. Default is \code{/imbs/external_data/annotation_and_references/annovar/humandb}.}

\item{build}{[\code{string}]\cr
Human genome build code. Default is 'hg19'.}

\item{exec}{[\code{string}]\cr
Path to \code{annotate_variation} script.}
}
\value{
Captured system output as \code{character} vector.
}
\description{
Annotate variant regions
}
\details{
Use ANNOVAR to find if variants hit exons or hit intergenic regions, or hit introns, or hit a non-coding RNA genes.
}
