% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_command.R
\name{check_command}
\alias{check_command}
\alias{assert_command}
\alias{test_command}
\alias{checkCommand}
\alias{assertCommand}
\alias{testCommand}
\title{Check if command exists}
\usage{
check_command(x)

assert_command(x, .var.name = vname(x), add = NULL)

test_command(x)

checkCommand(x)

assertCommand(x, .var.name = vname(x), add = NULL)

testCommand(x)
}
\arguments{
\item{x}{[\code{string}]\cr
Command to check.}

\item{.var.name}{[\code{string}]\cr
Name of the checked object to print in assertions. Defaults to the heuristic implemented in \code{\link{vname}}.}

\item{add}{[\code{\link[checkmate]{AssertCollection}}]\cr
Collection to store assertion messages. See \code{\link[checkmate]{AssertCollection}}.}
}
\value{
If the check is successful, i.e. the system command is found, returns \code{TRUE}. If the check is not successful, returns a string with the error message.
}
\description{
Check if command exists
}
\details{
See \code{\link[checkmate]{checkmate}} for details.
}
