% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_sets.R
\name{compare_sets}
\alias{compare_sets}
\title{Compare two vectors with simple set theory}
\usage{
compare_sets(a, b, label = NULL)
}
\arguments{
\item{a}{[vector(n)] First set of elements}

\item{b}{[vector(m)] Second set of elements}

\item{label}{[char(2)] Alternative labels for the sets. If NULL, names of 
\code{a} and \code{b} are taken.}
}
\value{
List with names of sets (\code{label}), size of sets (\code{n}) and
  actual content (\code{sets}).
}
\description{
Compare two vectors with simple set theory
}
\examples{

x <- 1:10
y <- 4:12
compare_sets(a = x, b = y)
}

