% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_paste.R
\name{convert_csv2csv}
\alias{convert_csv2csv}
\title{Changes delimiters in csv files}
\usage{
convert_csv2csv(in_file, out_file, from = c("de", "en"))
}
\arguments{
\item{in_file}{Input csv file(s) [character]}

\item{out_file}{Output csv file(s) [character]}

\item{from}{Direction}
}
\value{
Nothing, invoked for side effect of writing new files
}
\description{
Convert csv into csv2 files and vice versa. 
Can cope with vectors of input files
}

