% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.multirow.R
\name{do.multirow}
\alias{do.multirow}
\title{Combine rows with the same content}
\usage{
do.multirow(df, cols = 1:ncol(df), cmidrule = TRUE)
}
\arguments{
\item{df}{[\code{data.frame}]\cr
\code{data.frame} object of which to combine rows with 
same content to multirows.}

\item{cols}{[\code{integer}]\cr
Vector of column indices to apply multirow combination to.}

\item{cmidrule}{[\code{flag}]\cr
Should cmidrules inserted between groups of combinations?}
}
\value{
A \code{data.frame} with multirow(s) inserted.
}
\description{
Combine rows with the same content
}
\examples{
\dontrun{
df <- data.frame(FIRST = rep("A", 28), 
                 SECOND = c(rep("A", 4), rep ("B", 6), rep("A", 8), rep("B", 10)), 
                 THIRD = rep(c(LETTERS[0:13 \%\% 24 + 1]), each = 2), 
                 FORTH = LETTERS[0:27 \%\% 24 + 1])

print(
  xtable(
    do.multirow(df, cols = 1:3)), 
  booktabs = TRUE, 
  sanitize.text.function = identity, 
  include.rownames = FALSE
)
}
}
