% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{.plink_pca}
\alias{.plink_pca}
\title{Internal PCA computation with PLINK}
\usage{
.plink_pca(bed.file, bim.file, fam.file, output.prefix, num.evec, ...,
  exec, num.threads, memory)
}
\arguments{
\item{bed.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{bim.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{fam.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{output.prefix}{[\code{string}]\cr
The basename of the output files.}

\item{num.evec}{[\code{int}]\cr
Number of principal componentns to calculate.}

\item{...}{[\code{character}]\cr
Additional arguments passed to ALL PLINK calls (LD pruning, heterozygosity estimation, exclusion of samples).}

\item{exec}{[\code{string}]\cr
Path of PLINK executable.}

\item{num.threads}{[\code{int}]\cr
Number of CPUs usable by PLINK.
Default is determined by SLURM environment variables and at least 1.}

\item{memory}{[\code{int}]\cr
Memory for PLINK in Mb.
Default is determined by minimum of SLURM environment variables \code{SLURM_MEM_PER_CPU} and \code{num.threads * SLURM_MEM_PER_NODE} and at least 5000.}
}
\value{
Captured system outputs as \code{list} of \code{character} vectors and number of samples excluded per criteria.
}
\description{
Internal PCA computation with PLINK
}
