% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_genotypes.R
\name{harmonize_genotypes}
\alias{harmonize_genotypes}
\title{Genotype harmonization}
\usage{
harmonize_genotypes(input, ref, output, inputType, refType, outputType, ...,
  exec = "GenotypeHarmonizer")
}
\arguments{
\item{input}{[\code{string}]\cr
The base path of the data to align. The extensions are determined based on the input data type.}

\item{ref}{[\code{string}]\cr
The base path of the reference data used for alignment. The extensions are determined based on the input data type. If not specified the input data is simply converted to the specified output type.}

\item{output}{[\code{string}]\cr
The base path of the output data.}

\item{inputType}{[\code{string}]\cr
The input data type. If not defined will attempt to automatically select the first matching dataset on the specified path.}

\item{refType}{[\code{string}]\cr
The input data type. If not defined will attempt to automatically select the first matching dataset on the specified path.}

\item{outputType}{[\code{string}]\cr
The output data type. Defaults to --inputType.}

\item{...}{[\code{character}]\cr
Other options passed to \code{GenotypeHarmonizer}.}

\item{exec}{[\code{string}]\cr
Path of \code{GenotypeHarmonizer} executable.}
}
\value{
Captured system output as \code{character} vector.
}
\description{
Harmonization of genotype data stored using different file formats with different and potentially unknown strands. Linkage disequilibrium (LD) patterns are used to determine the correct strand GC and AT SNPs. This is a simple wrapper for \code{GenotypeHarmonizer}.
}
\details{
\url{https://github.com/molgenis/systemsgenetics/wiki/Genotype-Harmonizer}
}
