% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchtools.R
\name{load_or_create_registry}
\alias{load_or_create_registry}
\title{Load or create a \code{batchtools} registry}
\usage{
load_or_create_registry(file.dir, work.dir = getwd(), writeable = TRUE,
  overwrite = FALSE, ...)
}
\arguments{
\item{file.dir}{[\code{string}]\cr
File path where the registry is stored.}

\item{work.dir}{[\code{string}]\cr
File path where the working dir of the registry is set to.}

\item{overwrite}{[\code{logical}]\cr
Should an existing registry be removed and recreated?}

\item{...}{[any]\cr
Further arguments passed to \code{\link[batchtools]{makeRegistry}} and \code{\link[batchtools]{loadRegistry}}.}
}
\value{
A \code{\link[batchtools]{Registry}}.
}
\description{
Load or create a \code{batchtools} registry
}
