% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Create a manhattan plot}
\usage{
manhattan_plot(data, threshold, thin = FALSE,
  thin.param = list(logp.thin.thresh = 5, logp.thin.rate = 0.1, buffer =
  1e+06), header = list(chr = "CHR", pos = "BP", pval = "P"))
}
\arguments{
\item{data}{[\code{\link{data.frame}}]\cr
A \code{data.table} object containing at least 
information about chromosome, position and p value.}

\item{threshold}{[\code{number}]\cr
Single numeric to draw a threshold line at. On p value 
scale. Optional.}

\item{thin}{[\code{logical}]\cr
Should the data be thinned?}

\item{thin.param}{[\code{list}]\cr
A list with the following elements: 
\code{logp.thin.thresh} (\code{number}), 
\code{logp.thin.rate} (\code{number}) and 
\code{buffer} (\code{integer}) indicating at which rate 
points with a low log p value should be removed and how 
many base pairs at the start and end of each chromosome 
should not be excluded from thinning.}

\item{header}{[\code{list}]\cr
A list with the following elements:
\code{chr}, \code{pos} and \code{pval} (all 
\code{character}) indicating the columns to use for 
chromosome, position and p value information.}
}
\value{
A \code{\link[ggplot2]{ggplot2}} object.
}
\description{
Create a manhattan plot from a \code{data.frame}.
}
