% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrateInstalledPackages.R
\name{migrateInstalledPackages}
\alias{migrateInstalledPackages}
\title{Migrate installed packages}
\usage{
migrateInstalledPackages(old_dir)
}
\arguments{
\item{old_dir}{path to old package dir}
}
\value{
Vector of packages which are not on CRAN or Bioconductor
}
\description{
Uses Bioconductor to install R packages, skips already installed packages
}
\details{
Migrates installed packages from old version
}
