% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_paste.R
\name{named_file_path}
\alias{named_file_path}
\title{Named file path}
\usage{
named_file_path(..., sep = .Platform$file.sep)
}
\arguments{
\item{...}{character vectors.}

\item{sep}{the path separator to use.}
}
\value{
Named vector of concatenated file names.
}
\description{
Paste file names using named elements
}
\examples{
named_file_path("home", 
                c(homer = "projects", marge = "work"), 
                "doo")
}
