% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_paste.R
\name{named_paste}
\alias{named_paste}
\alias{named_paste0}
\title{Named paste}
\usage{
named_paste(..., sep = " ")

named_paste0(..., sep = "")
}
\arguments{
\item{...}{one or more R objects, to be converted to character vectors.}

\item{sep}{a character string to separate the items.}
}
\value{
Named vector of concatenated values.
}
\description{
Paste strings using named elements
}
\examples{
named_paste("a", c(b = "b", c = "c"))
}

