% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{plink_conversion}
\alias{plink_conversion}
\title{Convert VCF files to PLINK binary files}
\usage{
plink_conversion(seq.file, output.prefix, ..., exec = "plink",
  num.threads = max(1, as.integer(Sys.getenv("SLURM_NPROCS")), na.rm = TRUE),
  memory = max(5000, as.integer(Sys.getenv("SLURM_MEM_PER_CPU")) - 1000, na.rm
  = TRUE))
}
\arguments{
\item{seq.file}{[\code{string}]\cr
The VCF file path.}

\item{output.prefix}{[\code{string}]\cr
The basename of the new binary PLINK files.}

\item{...}{[\code{character}]\cr
Additional arguments passed to PLINK.}

\item{exec}{[\code{string}]\cr
Path of PLINK executable.}

\item{num.threads}{[\code{int}]\cr
Number of CPUs usable by PLINK.
Default is determined by SLURM environment variables and at least 1.}

\item{memory}{[\code{int}]\cr
Memory for PLINK in Mb.
Default is determined by SLURM environment variables and at least 5000.}
}
\value{
Captured system output as \code{character} vector.
}
\description{
Convert VCF files to PLINK binary files
}
\details{
See PLINK manual \url{https://www.cog-genomics.org/plink/1.9/}.
}
