% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{plink_fst}
\alias{plink_fst}
\title{Fst with PLINK}
\usage{
plink_fst(bfile, output.prefix, pop.file, outlier.removal = FALSE,
  outlier.sigma, num.outlier.evec, num.outlier.iter, ld.pruning.params,
  bed.file = NULL, bim.file = NULL, fam.file = NULL,
  exec = "plink2", num.threads, memory)
}
\arguments{
\item{bfile}{[\code{string}]\cr
The basename of the binary PLINK files.}

\item{output.prefix}{[\code{string}]\cr
The basename of the new binary PLINK files.}

\item{pop.file}{[\code{string}]\cr
File defining disjoint clusters/strata of samples. FIDs in the first column, IIDs in the second column, and cluster names in the third column.}

\item{outlier.removal}{[\code{flag}]\cr
Indicate if outliers shall be removed before final PCA computation.}

\item{outlier.sigma}{[\code{int}]\cr
How many standard deviations from the mean of a single PC is not an outlier?}

\item{num.outlier.evec}{[\code{int}]\cr
How many PCs shall be used for outlier detection?}

\item{num.outlier.iter}{[\code{int}]\cr
Number of outlier iterations.}

\item{ld.pruning.params}{[\code{list}]\cr
List with function arguments passed to \code{\link{plink_ld_pruning}}.}

\item{bed.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{bim.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{fam.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{exec}{[\code{string}]\cr
Path of PLINK executable.}

\item{num.threads}{[\code{int}]\cr
Number of CPUs usable by PLINK.
Default is determined by SLURM environment variables and at least 1.}

\item{memory}{[\code{int}]\cr
Memory for PLINK in Mb.
Default is determined by minimum of SLURM environment variables \code{SLURM_MEM_PER_CPU} and \code{num.threads * SLURM_MEM_PER_NODE} and at least 5000.}
}
\value{
A \code{list} of logs, Eigenvectors and Eigenvalues of the outlier removal and the final Fst values.
}
\description{
Fst with PLINK
}
\details{
First independent SNPs are extracted using PLINK LD pruning (\code{plink_ld_pruning}). If requested, outliers are detected and removed by calculating a PCA, calculating the mean (\eqn{\hat{\mu}_{j}}) and standard deviation (\eqn{\hat{\sigma}_{j}}) of each principal component, and excluding those samples who's PC value \eqn{v_{j}} exceeds \eqn{\hat{\mu}_{j}\pm\nu\,\hat{\sigma}_{j}} for at least one \eqn{j}, where \eqn{\nu} is \code{outlier.sigma}.
}
