% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{plink_ld_pruning}
\alias{plink_ld_pruning}
\title{LD pruning with PLINK}
\usage{
plink_ld_pruning(bfile, output.prefix, window.size, step.size, threshold, chr,
  ..., bed.file = NULL, bim.file = NULL, fam.file = NULL,
  exec = "plink", num.threads, memory)
}
\arguments{
\item{bfile}{[\code{string}]\cr
The basename of the binary PLINK files.}

\item{output.prefix}{[\code{string}]\cr
The basename of the new binary PLINK files.}

\item{window.size}{[\code{int}]\cr
Window size in kilobase.}

\item{step.size}{[\code{int}]\cr
Step size in kilobase.}

\item{threshold}{[\code{number}]\cr
Pairwise \eqn{r^2} threshold.}

\item{chr}{[\code{number}]\cr
Restrict LD pruning to specified chromosome.}

\item{...}{[\code{character}]\cr
Additional arguments passed to PLINK.}

\item{bed.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{bim.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{fam.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{exec}{[\code{string}]\cr
Path of PLINK executable.}

\item{num.threads}{[\code{int}]\cr
Number of CPUs usable by PLINK.
Default is determined by SLURM environment variables and at least 1.}

\item{memory}{[\code{int}]\cr
Memory for PLINK in Mb.
Default is determined by minimum of SLURM environment variables \code{SLURM_MEM_PER_CPU} and \code{num.threads * SLURM_MEM_PER_NODE} and at least 5000.}
}
\value{
Captured system output as \code{character} vector.
}
\description{
LD pruning with PLINK
}
\details{
See PLINK manual \url{https://www.cog-genomics.org/plink/1.9/}.
}
