% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{plink_logistic}
\alias{plink_logistic}
\title{GWA (single variant logistic regression) with PLINK}
\usage{
plink_logistic(bfile, output.prefix, test, sex, beta, intercept, ...,
  bed.file = NULL, bim.file = NULL, fam.file = NULL,
  exec = "plink2", num.threads, memory)
}
\arguments{
\item{bfile}{[\code{string}]\cr
The basename of the binary PLINK files.}

\item{output.prefix}{[\code{string}]\cr
The basename of the new binary PLINK files.}

\item{test}{[\code{string}]\cr
Genetic model to use. Choices are 'genotypic', 'hethom', 'dominant', and 'recessive'.
The 'genotypic' modifier adds an additive effect/dominance deviation 2df joint test (with two genotype-dependent variables in the regression, one with 0/1/2 coding and the second with 0/1/0 coding), while 'hethom' uses 0/0/1 and 0/1/0 coding instead. 
The 'dominant' and 'recessive' modifiers specify a model assuming full dominance or recessiveness, respectively, for the A1 allele.}

\item{sex}{[\code{flag}]\cr
Include sex as covariate?}

\item{beta}{[\code{flag}]\cr
Output regression coefficients instead of OR?}

\item{intercept}{[\code{flag}]\cr
Include intercept in output?}

\item{...}{[\code{character}]\cr
Additional arguments passed to PLINK call.}

\item{bed.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{bim.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{fam.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{exec}{[\code{string}]\cr
Path of PLINK executable.}

\item{num.threads}{[\code{int}]\cr
Number of CPUs usable by PLINK.
Default is determined by SLURM environment variables and at least 1.}

\item{memory}{[\code{int}]\cr
Memory for PLINK in Mb.
Default is determined by minimum of SLURM environment variables \code{SLURM_MEM_PER_CPU} and \code{num.threads * SLURM_MEM_PER_NODE} and at least 5000.}

\item{tmp.dir}{[\code{string}]\cr
Path where to save temporary files. If not set by user, defaults to \code{tempdir()}.}
}
\value{
A \code{list} of log (\code{gwa_log}, as \code{character}) and regression results (\code{gwa_result}, as \code{data.table}).
}
\description{
GWA (single variant logistic regression) with PLINK
}
\details{
Performs logistic regression given a case/control phenotype. For details see PLINK manual.
}
