% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{plink_marker_qc}
\alias{plink_marker_qc}
\title{Quality control on marker level with PLINK}
\usage{
plink_marker_qc(bfile, output.prefix, call.rate, min.maf, max.maf,
  hwe.pval, ..., bed.file = NULL, bim.file = NULL, fam.file = NULL,
  exec = "plink2", num.threads, memory)
}
\arguments{
\item{bfile}{[\code{string}]\cr
The basename of the binary PLINK files.}

\item{output.prefix}{[\code{string}]\cr
The basename of the new binary PLINK files.}

\item{call.rate}{[\code{double}]\cr
Filter out all variants with call rates falling below the provided threshold.}

\item{min.maf}{[\code{double}]\cr
Filter out all variants with minor allele frequency below the provided threshold.}

\item{max.maf}{[\code{double}]\cr#'
Filter out all variants with minor allele frequency above the provided threshold.}

\item{hwe.pval}{[\code{double}]\cr
Filter out all variants which have Hardy-Weinberg equilibrium exact test p-value below the provided threshold.}

\item{...}{[\code{character}]\cr
Additional arguments passed to PLINK.}

\item{bed.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{bim.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{fam.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{exec}{[\code{string}]\cr
Path of PLINK executable.}

\item{num.threads}{[\code{int}]\cr
Number of CPUs usable by PLINK.
Default is determined by SLURM environment variables and at least 1.}

\item{memory}{[\code{int}]\cr
Memory for PLINK in Mb.
Default is determined by minimum of SLURM environment variables \code{SLURM_MEM_PER_CPU} and \code{num.threads * SLURM_MEM_PER_NODE} and at least 5000.}
}
\value{
Captured system output as \code{character} vector.
}
\description{
Quality control on marker level with PLINK
}
\details{
See PLINK manual \url{https://www.cog-genomics.org/plink/1.9/}.
}
