% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{plink_merge}
\alias{plink_merge}
\title{Merge two PLINK datasets}
\usage{
plink_merge(old.prefix, new.prefix, merge.mode, output.prefix, ...,
  exec = "plink", num.threads = max(1,
  as.integer(Sys.getenv("SLURM_NPROCS")), na.rm = TRUE), memory = max(5000,
  as.integer(Sys.getenv("SLURM_MEM_PER_CPU")) - 1000, na.rm = TRUE))
}
\arguments{
\item{old.prefix}{[\code{string}]\cr
The basename of the first binary PLINK file set.}

\item{new.prefix}{[\code{string}]\cr
The basename of the second binary PLINK file set.}

\item{merge.mode}{[\code{int}]\cr
Merge mode.}

\item{output.prefix}{[\code{string}]\cr
The basename of the output binary PLINK file set.}

\item{...}{[\code{character}]\cr
Additional arguments passed to PLINK.}

\item{exec}{[\code{string}]\cr
Path of PLINK executable.}

\item{num.threads}{[\code{int}]\cr
Number of CPUs usable by PLINK.
Default is determined by SLURM environment variables and at least 1.}

\item{memory}{[\code{int}]\rc
Memory for PLINK in Mb.
Default is determined by SLURM environment variables and at least 5000.}
}
\value{
Captured system output as \code{character} vector.
}
\description{
Merge two PLINK datasets
}
\details{
See PLINK manual \url{https://www.cog-genomics.org/plink/1.9/}.
}
