% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{plink_merge_list}
\alias{plink_merge_list}
\title{Merge multiple PLINK datasets}
\usage{
plink_merge_list(bfile, output.prefix, merge.list, merge.mode, ...,
  bed.file = NULL, bim.file = NULL, fam.file = NULL, exec = "plink",
  num.threads = max(1, as.integer(Sys.getenv("SLURM_NPROCS")), na.rm = TRUE),
  memory = max(5000, as.integer(Sys.getenv("SLURM_MEM_PER_CPU")) - 1000, na.rm
  = TRUE))
}
\arguments{
\item{bfile}{[\code{string}]\cr
The basename of the binary PLINK files.}

\item{output.prefix}{[\code{string}]\cr
The basename of the new binary PLINK files.}

\item{merge.list}{[\code{string}]\cr
File listing other PLINK datasets to be merged.}

\item{merge.mode}{[\code{int}]\cr
Merge mode.}

\item{...}{[\code{character}]\cr
Additional arguments passed to PLINK.}

\item{bed.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{bim.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{fam.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{exec}{[\code{string}]\cr
Path of PLINK executable.}

\item{num.threads}{[\code{int}]\cr
Number of CPUs usable by PLINK.
Default is determined by SLURM environment variables and at least 1.}

\item{memory}{[\code{int}]\cr
Memory for PLINK in Mb.
Default is determined by SLURM environment variables and at least 5000.}
}
\value{
Captured system output as \code{character} vector.
}
\description{
Merge multiple PLINK datasets
}
\details{
See PLINK manual \url{https://www.cog-genomics.org/plink/1.9/}.
}
