% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{plink_normalize_vcf_conversion}
\alias{plink_normalize_vcf_conversion}
\title{Normalize and convert VCF files to PLINK binary files}
\usage{
plink_normalize_vcf_conversion(vcf.file, ref.file, output.prefix, ...,
  build = "b37", bcftools.exec = "bcftools", plink.exec = "plink2",
  num.threads, memory)
}
\arguments{
\item{vcf.file}{[\code{string}]\cr
The VCF file path.}

\item{ref.file}{[\code{string}]\cr
A human reference genome \code{fasta} file to normalize indels against.}

\item{output.prefix}{[\code{string}]\cr
The basename of the new binary PLINK files.}

\item{...}{[\code{character}]\cr
Additional arguments passed to PLINK.}

\item{build}{[\code{string}]\cr
Human genome build code to split the X chromosome into pseudo-autosomal region and pure X. Default is 'b37' (alias 'hg19').}

\item{bcftools.exec}{[\code{string}]\cr
Path of bcftools executable.}

\item{plink.exec}{[\code{string}]\cr
Path of PLINK executable.}

\item{num.threads}{[\code{int}]\cr
Number of CPUs usable by PLINK.
Default is determined by SLURM environment variables and at least 1.}

\item{memory}{[\code{int}]\cr
Memory for PLINK in Mb.
Default is determined by minimum of SLURM environment variables \code{SLURM_MEM_PER_CPU} and \code{num.threads * SLURM_MEM_PER_NODE} and at least 5000.}
}
\value{
Captured system output as \code{character} vector.
}
\description{
Normalize and convert VCF files to PLINK binary files
}
\details{
Based on the best practices on \url{http://apol1.blogspot.de/2014/11/best-practice-for-converting-vcf-files.html}. The procedure will take the VCF file, strip the variant IDs, split multi-allelic sites into bi-allelic sites, assign names to make sure indels will not become ambiguous, and finally convert to PLINK format. See PLINK manual \url{https://www.cog-genomics.org/plink/1.9/}.
}
