% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{plink_rm_long_indels}
\alias{plink_rm_long_indels}
\title{Remove very long INDELS}
\usage{
plink_rm_long_indels(bfile, output.prefix, max_length, ...,
  bed.file = NULL, bim.file = NULL, fam.file = NULL,
  exec = "plink2", num.threads, memory)
}
\arguments{
\item{bfile}{[\code{string}]\cr
The basename of the binary PLINK files.}

\item{output.prefix}{[\code{string}]\cr
The basename of the new binary PLINK files.}

\item{max_length}{[\code{integer}]\cr
The maximum length of any allele.}

\item{...}{[\code{character}]\cr
Additional arguments passed to PLINK.}

\item{bed.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{bim.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{fam.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{exec}{[\code{string}]\cr
Path of PLINK executable.}

\item{num.threads}{[\code{int}]\cr
Number of CPUs usable by PLINK.
Default is determined by SLURM environment variables and at least 1.}

\item{memory}{[\code{int}]\cr
Memory for PLINK in Mb.
Default is determined by minimum of SLURM environment variables \code{SLURM_MEM_PER_NODE} and \code{num.threads * SLURM_MEM_PER_CPU} and at least 5000.}
}
\value{
Captured system output as \code{character} vector.
}
\description{
Uses \code{cut}, \code{sort}, \code{uniq} and \code{awk} to find very long INDELS causing PLINK to be very memory hungry during analyses and excludes them using PLINK.
}
\details{
See PLINK manual \url{https://www.cog-genomics.org/plink/1.9/}.
}
