% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{plink_sample_qc}
\alias{plink_sample_qc}
\title{Quality control on sample level with PLINK}
\usage{
plink_sample_qc(bfile, output.prefix, call.rate, het.sigma,
  ld.pruning.params, ..., bed.file = NULL, bim.file = NULL,
  fam.file = NULL, exec = "plink2", num.threads, memory)
}
\arguments{
\item{bfile}{[\code{string}]\cr
The basename of the binary PLINK files.}

\item{output.prefix}{[\code{string}]\cr
The basename of the new binary PLINK files.}

\item{call.rate}{[\code{double}]\cr
Filter out all samples with call rates falling below the provided threshold.}

\item{het.sigma}{[\code{double}]\cr
Filter out all samples with mean heterozygosity rate above or below \code{het.sigma} times estimated heterozygosity standard deviation.}

\item{ld.pruning.params}{[\code{list}]\cr
List with function arguments passed to \code{\link{plink_ld_pruning}}.}

\item{...}{[\code{character}]\cr
Additional arguments passed to PLINK.}

\item{bed.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{bim.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{fam.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{exec}{[\code{string}]\cr
Path of PLINK executable.}

\item{num.threads}{[\code{int}]\cr
Number of CPUs usable by PLINK.
Default is determined by SLURM environment variables and at least 1.}

\item{memory}{[\code{int}]\cr
Memory for PLINK in Mb.
Default is determined by minimum of SLURM environment variables \code{SLURM_MEM_PER_CPU} and \code{num.threads * SLURM_MEM_PER_NODE} and at least 5000.}
}
\value{
Captured system output as \code{character} vector.
}
\description{
Quality control on sample level with PLINK
}
\details{
Heterozygosity estimation is not LD-sensitive, thus LD pruning is performend first using \code{\link{plink_ld_pruning}}. See PLINK manual \url{https://www.cog-genomics.org/plink/1.9/}.
}
