% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{plink_sex_imputation}
\alias{plink_sex_imputation}
\title{Sex imputation with PLINK}
\usage{
plink_sex_imputation(bfile, output.prefix, f.values, ..., bed.file = NULL,
  bim.file = NULL, fam.file = NULL, exec = "plink", num.threads, memory)
}
\arguments{
\item{bfile}{[\code{string}]\cr
The basename of the binary PLINK files.}

\item{output.prefix}{[\code{string}]\cr
The basename of the new binary PLINK files.}

\item{f.values}{[\code{numeric(2)}]\cr
\code{numeric} of length \code{2} with minimum F coefficients to impute females (\code{[1]}) or males (\code{[2]}).}

\item{...}{[\code{character}]\cr
Additional arguments passed to PLINK.}

\item{bed.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{bim.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{fam.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{exec}{[\code{string}]\cr
Path of PLINK executable.}

\item{num.threads}{[\code{int}]\cr
Number of CPUs usable by PLINK.
Default is determined by SLURM environment variables and at least 1.}

\item{memory}{[\code{int}]\cr
Memory for PLINK in Mb.
Default is determined by minimum of SLURM environment variables \code{SLURM_MEM_PER_CPU} and \code{num.threads * SLURM_MEM_PER_NODE} and at least 5000.}
}
\value{
Captured system output as \code{character} vector.
}
\description{
Sex imputation with PLINK
}
\details{
See PLINK manual \url{https://www.cog-genomics.org/plink/1.9/}.
}
