% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{plink_subset}
\alias{plink_subset}
\title{Subset samples and variants using PLINK}
\usage{
plink_subset(bfile, output.prefix, remove, keep, exclude, extract, ...,
  exec = "plink", num.threads = max(1,
  as.integer(Sys.getenv("SLURM_NPROCS")), na.rm = TRUE), memory = max(5000,
  as.integer(Sys.getenv("SLURM_MEM_PER_CPU")) - 1000, na.rm = TRUE))
}
\arguments{
\item{bfile}{[\code{string}]\cr
The basename of the binary PLINK files.}

\item{output.prefix}{[\code{string}]\cr
The basename of the new binary PLINK files.}

\item{remove}{[\code{string}]\cr
File name of a list of sample FIDs and IIDs to remove.}

\item{keep}{[\code{string}]\cr
File name of a list of sample FIDs and IIDs to keep.}

\item{exclude}{[\code{string}]\cr
File name of a list of variant names to remove.}

\item{extract}{[\code{string}]\cr
File name of a list of variant names to keep.}

\item{...}{[\code{character}]\cr
Additional arguments passed to PLINK.}

\item{exec}{[\code{string}]\cr
Path of PLINK executable.}

\item{num.threads}{[\code{int}]\cr
Number of CPUs usable by PLINK.
Default is determined by SLURM environment variables and at least 1.}

\item{memory}{[\code{int}]\cr
Memory for PLINK in Mb.
Default is determined by SLURM environment variables and at least 5000.}
}
\value{
Captured system output as \code{character} vector.
}
\description{
Subset samples and variants using PLINK
}
\details{
See PLINK manual \url{https://www.cog-genomics.org/plink/1.9/}.
}
