% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_3ptable.R
\name{print_3ptable}
\alias{print_3ptable}
\title{print_3ptable}
\usage{
print_3ptable(notes, ...)
}
\arguments{
\item{notes}{explanation of the table footnotes, see example}

\item{...}{everything you would give to print.xtable}
}
\value{
a well formated Latex-table
}
\description{
print xtable objekts with table notes using threeparttable
}
\examples{
\dontrun{
notes <- paste('\\\\begin{tablenotes}',
               '\\\\item [1] Das ist der Parameterschätzer $\\\\beta$',
               '\\\\item [2] Das ist der p-Wert',
               '\\\\end{tablenotes}',
               sep = '\\n')
               
print_3ptable(notes = notes, xtable(data2, caption = "Regression Coefficients",
                                    label = "tab:3", digits = c(0, 4, 4, 2, 2),
                                    display = c('s', 'f','f','f','e'), ),
              include.rownames = TRUE)
              }
}

