% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character-vectors.R
\name{qw}
\alias{qw}
\title{Emulate Perl's \code{qw} (quoted words) function}
\usage{
qw(...)
}
\arguments{
\item{...}{Objects to be treated as characters}
}
\value{
Character vector
}
\description{
Hadley's solution on stackoverflow:
\url{http://stackoverflow.com/questions/520810/does-r-have-quote-like-operators-like-perls-qw}
}
\examples{
qw(a, b, c)
qw(1, a, 4)

}

