% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.sample.R
\name{read.sample}
\alias{read.sample}
\title{Read SNPTEST sample file}
\usage{
read.sample(file)
}
\arguments{
\item{file}{Name of the sample file}
}
\value{
data.frame containing the data
}
\description{
Reads SNPTEST sample files, ignores the second line though
}

