% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_call.R
\name{system_call}
\alias{system_call}
\title{Wrapper for \code{system2}}
\usage{
system_call(bin, ...)
}
\arguments{
\item{bin}{[\code{string}]\cr
The system command to be invoked as character string.}

\item{...}{[any]\cr
Additional arguments passed to \code{system2}.}
}
\value{
The captured output as a string.
}
\description{
Wrapper for \code{system2}
}
\details{
Output of '\code{stdout}' and '\code{stderr}' is captured in a character vector and checked for the exit status. If the status is not \code{0} an error is thrown. Otherwise the captured output is returned as a string.
}
