% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchjobs-system2.R
\name{update_system2_job}
\alias{update_system2_job}
\title{update_system2_job}
\usage{
update_system2_job(reg_name, sys_cmd, vec_cmd_args, more_cmd_args,
  env = character(), update = FALSE)
}
\arguments{
\item{reg_name}{[char] Name of registry to load}

\item{sys_cmd}{[char] Name of script to run through \code{\link{system2()}}}

\item{vec_cmd_args}{[char] Vector to loop over}

\item{more_cmd_args}{[char] Static arguments, handed to more.args of \code{\link{BatchJobs::batchMap}}}

\item{env}{[char] Directly passed through to  \code{\link{system2()}}}

\item{update}{[logical] TRUE deletes the registry to force re-execution of jobs, default is FALSE}
}
\value{
Nothing, throws an error if not all jobs are finished
}
\description{
Wrapper around typically batchMap() setup
}

