% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character-vectors.R
\name{verbatimize}
\alias{verbatimize}
\title{Verbatimize string}
\usage{
verbatimize(x, sep = ", ")
}
\arguments{
\item{x}{Character vector to be encapsulated}

\item{sep}{String; How pasted elements should be separated, Default is ", ".}
}
\value{
String of pasted values
}
\description{
Encapsulate elements of vector in LaTeX verbatim commands
}
