% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{write_plink_ids}
\alias{write_plink_ids}
\title{Write SNP or sample IDs to file}
\usage{
write_plink_ids(ids, file, sample = TRUE)
}
\arguments{
\item{ids}{Vector of SNP or Sample IDs}

\item{file}{File name to write to}

\item{sample}{Logical. Should be TRUE for samples and FALSE for SNPs}
}
\description{
Write SNP or sample IDs to file
}

