% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{plink_merge}
\alias{plink_merge}
\title{Merge two PLINK datasets}
\usage{
plink_merge(first.prefix, second.prefix, merge.mode, output.prefix, ...,
  first.bed.file = NULL, first.bim.file = NULL, first.fam.file = NULL,
  second.bed.file = NULL, second.bim.file = NULL, second.fam.file = NULL,
  exec = "plink", num.threads = max(1,
  as.integer(Sys.getenv("SLURM_NPROCS")), na.rm = TRUE), memory = max(5000,
  min(as.integer(Sys.getenv("SLURM_MEM_PER_NODE")) - 1000, num.threads *
  as.integer(Sys.getenv("SLURM_MEM_PER_CPU")) - 1000, na.rm = TRUE), na.rm =
  TRUE))
}
\arguments{
\item{first.prefix}{[\code{string}]\cr
The basename of the first binary PLINK file set.}

\item{second.prefix}{[\code{string}]\cr
The basename of the second binary PLINK file set.}

\item{merge.mode}{[\code{int}]\cr
Merge mode.}

\item{output.prefix}{[\code{string}]\cr
The basename of the output binary PLINK file set.}

\item{...}{[\code{character}]\cr
Additional arguments passed to PLINK.}

\item{first.bed.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{first.bim.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{first.fam.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{second.bed.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{second.bim.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{second.fam.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{exec}{[\code{string}]\cr
Path of PLINK executable.}

\item{num.threads}{[\code{int}]\cr
Number of CPUs usable by PLINK.
Default is determined by SLURM environment variables and at least 1.}

\item{memory}{[\code{int}]\cr
Memory for PLINK in Mb.
Default is determined by minimum of SLURM environment variables \code{SLURM_MEM_PER_CPU} and \code{num.threads * SLURM_MEM_PER_NODE} and at least 5000.}
}
\value{
Captured system output as \code{character} vector.
}
\description{
Merge two PLINK datasets
}
\details{
See PLINK manual \url{https://www.cog-genomics.org/plink/1.9/}.
}
