% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{plink_sex_imputation}
\alias{plink_sex_imputation}
\title{Sex imputation with PLINK}
\usage{
plink_sex_imputation(bfile, output.prefix, f.values, ..., exec = "plink",
  num.threads = max(1, as.integer(Sys.getenv("SLURM_NPROCS")), na.rm = TRUE),
  memory = max(5000, as.integer(Sys.getenv("SLURM_MEM_PER_CPU")) - 1000, na.rm
  = TRUE))
}
\arguments{
\item{bfile}{[\code{string}]\cr
The basename of the binary PLINK files.}

\item{output.prefix}{[\code{string}]\cr
The basename of the new binary PLINK files.}

\item{f.values}{[\code{numeric(2)}]\cr
\code{numeric} of length \code{2} with minimum F coefficients to impute females (\code{[1]}) or males (\code{[2]}).}

\item{...}{[\code{character}]\cr
Additional arguments passed to PLINK.}

\item{exec}{[\code{string}]\cr
Path of PLINK executable.}

\item{num.threads}{[\code{int}]\cr
Number of CPUs usable by PLINK.
Default is determined by SLURM environment variables and at least 1.}

\item{memory}{[\code{int}]\cr
Memory for PLINK in Mb.
Default is determined by SLURM environment variables and at least 5000.}
}
\value{
Captured system output as \code{character} vector.
}
\description{
Sex imputation with PLINK
}
\details{
See PLINK manual \url{https://www.cog-genomics.org/plink/1.9/}.
}
