% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{plink_subset}
\alias{plink_subset}
\title{Subset samples and variants using PLINK}
\usage{
plink_subset(bfile, output.prefix, remove, keep, exclude, extract, ...,
  bed.file = NULL, bim.file = NULL, fam.file = NULL,
  exec = "plink2", num.threads, memory)
}
\arguments{
\item{bfile}{[\code{string}]\cr
The basename of the binary PLINK files.}

\item{output.prefix}{[\code{string}]\cr
The basename of the new binary PLINK files.}

\item{remove}{[\code{string}]\cr
File name of a list of sample FIDs and IIDs to remove.}

\item{keep}{[\code{string}]\cr
File name of a list of sample FIDs and IIDs to keep.}

\item{exclude}{[\code{string}]\cr
File name of a list of variant names to remove.}

\item{extract}{[\code{string}]\cr
File name of a list of variant names to keep.}

\item{...}{[\code{character}]\cr
Additional arguments passed to PLINK.}

\item{bed.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{bim.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{fam.file}{[\code{string}]\cr
Alternative to \code{bfile} interface. Specify \code{bed}, \code{bim} and \code{fam} files individually.}

\item{exec}{[\code{string}]\cr
Path of PLINK executable.}

\item{num.threads}{[\code{int}]\cr
Number of CPUs usable by PLINK.
Default is determined by SLURM environment variables and at least 1.}

\item{memory}{[\code{int}]\cr
Memory for PLINK in Mb.
Default is determined by minimum of SLURM environment variables \code{SLURM_MEM_PER_CPU} and \code{num.threads * SLURM_MEM_PER_NODE} and at least 5000.}
}
\value{
Captured system output as \code{character} vector.
}
\description{
Subset samples and variants using PLINK
}
\details{
See PLINK manual \url{https://www.cog-genomics.org/plink/1.9/}.
}
