% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{write_plink_ids}
\alias{write_plink_ids}
\title{Write SNP or sample IDs to file}
\usage{
write_plink_ids(ids, file, update, sample = TRUE)
}
\arguments{
\item{ids}{Vector of SNP or Sample IDs}

\item{file}{File name to write to}

\item{update}{Logical, Default is FALSE. Skip writing file if it exists}

\item{sample}{Logical. Should be TRUE for samples (the default) and FALSE for
SNPs}
}
\description{
Write SNP or sample IDs to file
}
